requirements.txt
Pillow
fastapi
reportlab
google-genai
python-dotenv
uvicorn
python-multipart
